refactor: moved the flow to another file
