+++ frontend/src/pages/OnboardingPageV2/OnboardingPageV2.styles.scss
@@ -72,12 +72,6 @@
 		margin-bottom: 24px !important;
 	}
 
-	&__radio-buttons {
-		// display: flex;
-		// flex-wrap: wrap;
-		// gap: 12px;
-	}
-
 	&__radio-label {
 		display: flex;
 		cursor: pointer;
