+++ frontend/src/pages/OnboardingPageV2/QuestionBlock.tsx
@@ -2,7 +2,7 @@ import { SearchOutlined } from '@ant-design/icons';
 import { Avatar, Button, Flex, Input, Typography } from 'antd';
 import React, { useState } from 'react';
 
-import { Question } from './OnboardingPageV2'; // Adjust the import path as necessary
+import { Question } from './OnboardingAddDataSource'; // Adjust the import path as necessary
 
 const { Title, Paragraph, Text } = Typography;
